+++ src/pages/login/LoginComponent.css
@@ -259,30 +259,3 @@ form {
     font-size: 0.9em;
     text-align: center;
 }
-
-/* 알림창 */
-.btn-alert-close {
-    width: 100px;
-    padding: 10px;
-    background-color: #f0c040;
-    border: none;
-    border-radius: 5px;
-    color: #fff;
-    font-size: 14px;
-    font-weight: bold;
-    cursor: pointer;
-    margin-top: 20px;
-}
-
-.alert-h2 {
-    margin-top: 5px;
-}
-
-.alert-p { 
-    margin-top:10px;
-
-}
-
-.btn-alert-close:hover {
-    background-color: #e0a030;
-}
