+++ src/pages/mypage/UpdateCustomerComponent.css
@@ -154,33 +154,6 @@ label {
     font-size: 12px;
 }
 
-/* 알림창 */
-.btn-alert-close {
-    width: 100px;
-    padding: 10px;
-    background-color: #f0c040;
-    border: none;
-    border-radius: 5px;
-    color: #fff;
-    font-size: 14px;
-    font-weight: bold;
-    cursor: pointer;
-    margin-top: 20px;
-}
-
-.alert-h2 {
-    margin-top: 5px;
-}
-
-.alert-p { 
-    margin-top:10px;
-
-}
-
-.btn-alert-close:hover {
-    background-color: #e0a030;
-}
-
 .pw-match {
     color: green;
     font-size: 13px;
